# gateway/docs/swagger.yaml
basePath: /
definitions:
  internal_handler.CreateDriverRequest:
    properties:
      carBrand:
        example: Toyota
        type: string
      carModel:
        example: Corolla
        type: string
      firstName:
        example: Ahmet
        type: string
      lastName:
        example: Demir
        type: string
      lat:
        example: 41.0431
        type: number
      lon:
        example: 29.0099
        type: number
      plate:
        example: 34ABC123
        type: string
      taksiType:
        enum:
        - sari
        - turkuaz
        - siyah
        example: sari
        type: string
    required:
    - carBrand
    - carModel
    - firstName
    - lastName
    - lat
    - lon
    - plate
    - taksiType
    type: object
  internal_handler.Driver:
    properties:
      carBrand:
        type: string
      carModel:
        type: string
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      location:
        properties:
          lat:
            type: number
          lon:
            type: number
        type: object
      plate:
        type: string
      taxiType:
        type: string
      updatedAt:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_handler.ListDriversResponse:
    properties:
      drivers:
        items:
          $ref: '#/definitions/internal_handler.Driver'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
  internal_handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_handler.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_handler.NearbyDriverResponse:
    properties:
      distanceKm:
        type: number
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      plate:
        type: string
      taxiType:
        type: string
    type: object
  internal_handler.UpdateDriverRequest:
    properties:
      carBrand:
        example: Honda
        type: string
      carModel:
        example: Civic
        type: string
      firstName:
        example: Mehmet
        type: string
      lastName:
        example: Kurt
        type: string
      lat:
        example: 41.0082
        type: number
      lon:
        example: 28.9784
        type: number
      plate:
        example: 34XYZ789
        type: string
      taksiType:
        enum:
        - sari
        - turkuaz
        - siyah
        example: turkuaz
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@bitaksi.com
    name: API Support
  description: TaxiHub API Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gateway API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/internal_handler.LoginResponse'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Login
      tags:
      - auth
  /drivers:
    get:
      description: Get a paginated list of drivers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of drivers
          schema:
            $ref: '#/definitions/internal_handler.ListDriversResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List drivers
      tags:
      - drivers
    post:
      consumes:
      - application/json
      description: Create a new taxi driver
      parameters:
      - description: Driver information
        in: body
        name: driver
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateDriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Driver created successfully
          schema:
            $ref: '#/definitions/internal_handler.Driver'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new driver
      tags:
      - drivers
  /drivers/{id}:
    get:
      description: Get driver details by ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Driver details
          schema:
            $ref: '#/definitions/internal_handler.Driver'
        "404":
          description: Driver not found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get a driver by ID
      tags:
      - drivers
    put:
      consumes:
      - application/json
      description: Update an existing driver
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver update information
        in: body
        name: driver
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Driver updated successfully
          schema:
            $ref: '#/definitions/internal_handler.Driver'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Driver not found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a driver
      tags:
      - drivers
  /drivers/nearby:
    get:
      description: Find drivers within 6km radius
      parameters:
      - description: Latitude
        format: float64
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        format: float64
        in: query
        name: lon
        required: true
        type: number
      - description: Taxi type (sari, turkuaz, siyah)
        in: query
        name: taksiType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of nearby drivers sorted by distance
          schema:
            items:
              $ref: '#/definitions/internal_handler.NearbyDriverResponse'
            type: array
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Find nearby drivers
      tags:
      - drivers
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token with "Bearer " prefix (e.g., "Bearer eyJhbGci...").
      IMPORTANT: You must include "Bearer " before the token.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"